requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pyrealsense2>=2.54.0
open3d>=0.18.0
ultralytics>=8.0.0  # For YOLOv8
pyyaml>=6.0  # For config
pytest>=7.4  # For tests